Create ToDo class
The ToDo class contains all the attributes to describe a ToDo instance. A ToDo instance can be a one-off task, a task that needs to be completed multiple times, or a task that is time-based. The ToDo instance can be recurring over various periods of time. The class also allows the user to specify the importance of a ToDo item.