Add functionality to add a new ToDo
The user can now enter information about a new ToDo item to add it to their list of ToDo's. new_todo.dart functions as a wrapper that shows different child widgets to get information from the user. The new_todo screen has been added to the list of routes, and can be accessed by clicking the floating action button on the ToDos screen.