Fix bug of recurTimes not being in the range of NumberPicker
There was a bug where the user could say that the task was not recurring, which sets recurTimes to 0. If the user went back and said that the task was recurring, recurTimes would not be in the range of the NumberPicker. Setting recurTimes to null when going back resolves this bug.