Add BottomSheet when a ToDo item is pressed
A BottomSheet now appears when a user taps a ToDo item. The sheet functionality has not been implemented.

The BottomSheet will be used to complete one-off tasks, decrement the number of times remaining to complete a task to be done multiple times, or start a timer to reduce the remaining duration of a time-based task.

Function names in the ToDo service have been updated to be consistent with "to-do" being two words.