Implement BottomSheet menu for duration-based tasks
When the user taps on a task that is to be done for a certain period of time, they can indicate how much of that duration they have done the task for. The widget decrements the durationRemaining of the ToDo instance. If the durationRemaining is 0, the completed status of the ToDo instance is set to true.