Implement basic ToDo editing functionality
The ToDos can now be edited by tapping the three dots icon on a ToDo tile. Users can edit the task String, and the number of times to do the task or duration, depending on the nature of the task.

TODO: User will be able to change the recurring status of the ToDo, the due date, and the importance.