Implement recurring funcionality of ToDos
For recurring ToDos, if the most recent dueDate has passed, the dueDate will be updated according to the recurWindow. For example, if the first dueDate was 1/14 and the recurWindow was "weekly", the dueDate will be updated to 1/21, once 1/14 has passed. The ToDo will be set to incomplete and the timesRemaining or durationRemaining will be set to the numTimes or duration of the ToDo, respectively.