Fix bugs with creating a new ToDo
When the user selected that a task was recurring and then specified the recur window, this recur window persisted even if the user went back and said that the task was not recurring. This has been fixed.

In addition, the task and importance now persist. If the user enters the task, continues, and then goes back to enter the task, the original task is still there. This allows them to edit the task instead of starting from scratch.