Add explanations throughout the app to improve usability
The NoToDo screen points the user to the floating action button, making it more clear how to create a new ToDo.

Instructions have been added throughout the process of creating a ToDo to make it easier to understand.

The sign-in/register screen now has hint text that makes it clear that the top TextFormField is for an email, and the bottom TextFormField is for a password.