Stylize sign-in and register screens
The authenticate widget is now a wrapper for two separate widgets for signing in and registering. There is now an animation between the sign-in and register widgets.

The colors of the widgets are set according to a theme in the constants.dart file. The current theme has a dark blue as the main color, with light blue and bright green accents.