Override back button functionality when creating a new ToDo
Since all the screens for creating a new ToDo are on the same route, hitting the back button on an Android would go back to the ToDos screen, instead of going back to the previous screen in the process of creating a new ToDo. Now the back button works as expected.